Improve CI: robust vcpkg integration and unified cross-platform CMake workflow (#173)
* WIP on master

* Pipeline updates for Ubuntu

* Pipeline updates

* Updated setup and cache for vcpkg.

* Added vcpkg builtin-baseline.

* Removed vcpkg binary caching.

* Working on pipeline fixes.

* Updates vcpkg action to fix "action not found"

Updates the vcpkg action version to resolve the "action not found" error in the pipeline.
The configuration is also updated to specify the manifest directory, target triplet, and provide the GitHub token.

* Updates builtin-baseline

Updates the builtin-baseline in vcpkg.json to align with the latest changes in the vcpkg registry.

* Adds fetch-depth to vcpkg action

Adds fetch-depth: 0 to the vcpkg action to ensure the full history is available for vcpkg to function correctly.
This resolves potential issues where vcpkg might fail due to shallow clones.

* Adds bash shell to CMake workflow steps

Specifies the bash shell for the Configure, Build, and Run tests steps in the CMake multi-platform workflow.

This ensures consistency and avoids potential issues arising from relying on the default shell.

* Removes unnecessary line break in CMake command

Removes a line break in the CMake configure command within the GitHub Actions workflow.
This change simplifies the command's structure without affecting its functionality.

* Configures CMake separately for each OS

Configures CMake differently based on the operating system to ensure proper build setup. This change separates the configuration steps for Linux and Windows. The Windows configuration explicitly sets the toolchain file, target triplet and disables manifest mode.

* Uses vcpkg toolchain file from vcpkg step

The VCPKG toolchain file path is now dynamically obtained from the vcpkg step output, instead of being hardcoded.

This makes the build process more robust and less prone to errors related to incorrect or outdated paths.

---------

Co-authored-by: John Farrier <john.farrier@polyrhythm.com>